arduinoslave
Added the code that I tested earlier with attempting to keep the LED on for a set amount of time. This system may not work as I am still trying to figure out the delays so dont expect the LED to stay on for the time we set in the code.